Learned how to and moved current counter function to seperate file to run out of App.js + changed initial value to 10
