+++ Client-App/package.json
@@ -10,10 +10,11 @@
     "test": "jest"
   },
   "dependencies": {
-    "expo": "~53.0.9",
+    "expo": "53.0.13",
     "expo-status-bar": "~2.2.3",
     "react": "19.0.0",
-    "react-native": "0.79.2"
+    "react-native": "0.79.4",
+    "jest-expo": "53.0.7"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
